Add core Router and Coordinator contracts
 - Introduces the Router as the main entry point for managing subscriptions and routing requests.
 - Adds the Coordinator to orchestrate the lifecycle of individual computation requests.
 - Establishes a clear separation of concerns for a modular and scalable architecture.

+++ src/v1_0_0/consumer/BaseConsumer.sol
@@ -91,7 +91,7 @@ abstract contract BaseConsumer is Routable {
     /// @param containerId if lazy subscription, subscription compute container ID, else empty
     /// @param index if lazy subscription, `Inbox` lazy store index, else empty
     function rawReceiveCompute(
-        uint32 subscriptionId,
+        uint64 subscriptionId,
         uint32 interval,
         uint16 redundancy,
         address node,
